move invertAllCategories into VisibilityUtils
